Changed example to an actual example isntead of a test
